+++ frog-utils/src/index.js
@@ -1,8 +1,8 @@
 // @flow
-import React, { Component } from 'react';
+import React from 'react';
 
 import { compose, withHandlers, withState } from 'recompose';
-import { shuffle, omit } from 'lodash';
+import { shuffle } from 'lodash';
 
 export {
   default as EnhancedForm,
@@ -11,6 +11,7 @@ export {
 } from './EnhancedForm';
 export { generateReactiveFn, inMemoryReactive } from './generateReactiveFn';
 export { Highlight } from './highlightSubstring';
+export { ReactiveText } from './ReactiveText';
 export { msToString } from './msToString';
 export { default as uuid } from 'cuid';
 export { default as colorRange } from './colorRange';
@@ -58,7 +59,7 @@ export {
 } from './ActivityComponents/TableView';
 export { default as TreeView } from './ActivityComponents/TreeView';
 
-export const A = ({ onClick, children, ...rest }: any): any => (
+export const A = ({ onClick, children, ...rest }: any): any =>
   <a
     href="#"
     onClick={e => {
@@ -68,8 +69,7 @@ export const A = ({ onClick, children, ...rest }: any): any => (
     {...rest}
   >
     {children}
-  </a>
-);
+  </a>;
 
 export const currentDate = (): string => {
   const d = new Date();
@@ -137,65 +137,7 @@ export const wordWrap = (text: string, maxLength: number): string[] => {
 };
 
 const groupchars = 'ABCDEFGHIJKLMNOPQRSTUWXYZ123456789'.split('');
-export const getSlug = (n: number) =>
-  shuffle(groupchars)
-    .slice(0, n)
-    .join('');
-
-type ReactivePropsT = {
-  path: string | string[],
-  dataFn: Object,
-  type: 'textarea' | 'textinput'
-};
-
-export class ReactiveText extends Component {
-  textRef: any;
-  binding: any;
-  state: ReactivePropsT;
-
-  update = (props: ReactivePropsT) => {
-    this.setState({ path: props.path, dataFn: props.dataFn });
-    if (this.binding) {
-      this.binding.destroy();
-    }
-    this.binding = props.dataFn.bindTextField(this.textRef, props.path);
-  };
-
-  componentDidMount() {
-    this.update(this.props);
-  }
-
-  componentWillReceiveProps(nextProps: ReactivePropsT) {
-    if (
-      (nextProps.dataFn && nextProps.dataFn.doc.id) !==
-        (this.props.dataFn && this.props.dataFn.doc.id) ||
-      this.props.path !== nextProps.path ||
-      this.props.type !== nextProps.type
-    ) {
-      this.update(nextProps);
-    }
-  }
-
-  componentWillUnmount() {
-    if (this.binding) {
-      this.binding.destroy();
-    }
-  }
-
-  render() {
-    const rest = omit(this.props, ['path', 'dataFn']);
-    return this.props.type === 'textarea' ? (
-      <textarea ref={ref => (this.textRef = ref)} {...rest} defaultValue="" />
-    ) : (
-      <input
-        type="text"
-        ref={ref => (this.textRef = ref)}
-        {...rest}
-        defaultValue=""
-      />
-    );
-  }
-}
+export const getSlug = (n: number) => shuffle(groupchars).slice(0, n).join('');
 
 // If you try to insert value=0 path=['a', 'b', 'c']
 // into sharedb with doc={ 'a': { d:5 } }
